# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BizPulse API — AI-Powered Business Identity and Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ComplianceStatusStatus =
  (typeof ComplianceStatusStatus)[keyof typeof ComplianceStatusStatus];

export const ComplianceStatusStatus = {
  compliant: "compliant",
  warning: "warning",
  non_compliant: "non_compliant",
  pending: "pending",
} as const;

export interface ComplianceStatus {
  department: string;
  status: ComplianceStatusStatus;
  expiresAt?: string | null;
  note?: string | null;
}

export type BusinessStatus =
  (typeof BusinessStatus)[keyof typeof BusinessStatus];

export const BusinessStatus = {
  Active: "Active",
  Action_Required: "Action Required",
  "Non-Compliant": "Non-Compliant",
} as const;

export interface Business {
  ubid: string;
  name: string;
  owner: string;
  city: string;
  sector: string;
  gstin?: string | null;
  udyam?: string | null;
  municipalLicence?: string | null;
  fssai?: string | null;
  complianceScore: number;
  turnover: number[];
  status: BusinessStatus;
  alerts: string[];
  compliance: ComplianceStatus[];
  directors?: string[];
  address?: string;
  registrationDate?: string;
}

export type IntelligenceAlertSeverity =
  (typeof IntelligenceAlertSeverity)[keyof typeof IntelligenceAlertSeverity];

export const IntelligenceAlertSeverity = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface IntelligenceAlert {
  id: string;
  severity: IntelligenceAlertSeverity;
  category: string;
  title: string;
  description: string;
  district?: string | null;
  sector?: string | null;
  affectedCount?: number | null;
  createdAt: string;
}

export interface RawRecord {
  source: string;
  name: string;
  pan?: string | null;
  address?: string | null;
}

export interface ResolveRequest {
  records: RawRecord[];
}

export interface MatchScore {
  fuzzyName: number;
  panCrossRef: number;
  addressSimilarity: number;
  overall: number;
}

export interface ResolvedEntity {
  ubid: string;
  canonicalName: string;
  matchedRecords: RawRecord[];
  scores: MatchScore;
}

export interface ResolveResponse {
  entities: ResolvedEntity[];
  totalInputRecords: number;
  totalResolved: number;
  processingTimeMs: number;
}

export interface PlatformStats {
  totalBusinesses: number;
  unifiedToday: number;
  complianceRate: number;
  activeAlerts: number;
  districtsActive: number;
  sectorsMonitored: number;
}

export interface SectorAnalytic {
  district: string;
  sector: string;
  complianceRate: number;
  totalBusinesses: number;
  compliantCount: number;
}

export type ListBusinessesParams = {
  sector?: string;
  city?: string;
  status?: string;
  search?: string;
};