# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BizPulse API — AI-Powered Business Identity and Intelligence Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses:
    get:
      operationId: listBusinesses
      tags: [businesses]
      summary: List all businesses
      description: Returns all registered businesses with their UBID profiles
      parameters:
        - name: sector
          in: query
          required: false
          schema:
            type: string
        - name: city
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Business"
  /businesses/{ubid}:
    get:
      operationId: getBusiness
      tags: [businesses]
      summary: Get business by UBID
      description: Returns full 360° profile for a business
      parameters:
        - name: ubid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: Business not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    ComplianceStatus:
      type: object
      properties:
        department:
          type: string
        status:
          type: string
          enum: [compliant, warning, non_compliant, pending]
        expiresAt:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
      required:
        - department
        - status
    Business:
      type: object
      properties:
        ubid:
          type: string
        name:
          type: string
        owner:
          type: string
        city:
          type: string
        sector:
          type: string
        gstin:
          type: string
          nullable: true
        udyam:
          type: string
          nullable: true
        municipalLicence:
          type: string
          nullable: true
        fssai:
          type: string
          nullable: true
        complianceScore:
          type: integer
        turnover:
          type: array
          items:
            type: number
        status:
          type: string
          enum: [Active, "Action Required", Non-Compliant]
        alerts:
          type: array
          items:
            type: string
        compliance:
          type: array
          items:
            $ref: "#/components/schemas/ComplianceStatus"
        directors:
          type: array
          items:
            type: string
        address:
          type: string
        registrationDate:
          type: string
      required:
        - ubid
        - name
        - owner
        - city
        - sector
        - complianceScore
        - turnover
        - status
        - alerts
        - compliance
...