# lib/api-zod/src/generated/types/business.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BizPulse API — AI-Powered Business Identity and Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { BusinessStatus } from "./businessStatus";
import type { ComplianceStatus } from "./complianceStatus";

export interface Business {
  ubid: string;
  name: string;
  owner: string;
  city: string;
  sector: string;
  gstin?: string | null;
  udyam?: string | null;
  municipalLicence?: string | null;
  fssai?: string | null;
  complianceScore: number;
  turnover: number[];
  status: BusinessStatus;
  alerts: string[];
  compliance: ComplianceStatus[];
  directors?: string[];
  address?: string;
  registrationDate?: string;
}