# lib/api-zod/src/generated/types/complianceStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BizPulse API — AI-Powered Business Identity and Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { ComplianceStatusStatus } from "./complianceStatusStatus";

export interface ComplianceStatus {
  department: string;
  status: ComplianceStatusStatus;
  expiresAt?: string | null;
  note?: string | null;
}